Fix compass no network issue (#716)
* Update index.js

* w

* fixes

* better

* remove not used

* Revert "remove not used"

This reverts commit 21e7539457a27304f3406ea061e97a1b657582d4.

* Revert "better"

This reverts commit b58456a122d75d76920da09ad4a0e2fd31889410.

* Revert "fixes"

This reverts commit 1637a551ee60eb069aca3bcb0fe5e1215e253610.

* Revert "w"

This reverts commit 039020a1a0723540fdcc67a504082f8bf35d57eb.

* Revert "Update index.js"

This reverts commit d4cdf8f6057d425ea4ea8d44f036c6a405ed2976.

* Aut 13147 b (#717)

* wip

* Update index.js

* remove check

* refactor

* Update index.tsx

* Update index.js

* Revert "Update index.tsx"

This reverts commit 965c25e1d1f0e41e15d4723c6c0c77bc73458ae0.

* revert

* Revert "Revert "Update index.tsx""

This reverts commit ca45c95b301c960c52209b8742c9a78665f550ff.

* Update index.js

* Update index.js
